dspy/home_healthcare_remedy_advisor/requirements.txt
chromadb
python-dotenv
dspy-ai
BeautifulSoup4
langchain
genai
google-generativea
sentence-transformers
